Move ui components to logical folders
